requirements.txt
streamlit
pdfplumber
pyMuPDF
torch
torchvision
pillow
scikit-learn
